# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusCare issue reporting and resolution API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Issue {
  id: number;
  issueId: string;
  description: string;
  shortDescription: string;
  category: string;
  aiCategory: string;
  priority: string;
  aiPriority: string;
  priorityReason: string;
  status: string;
  campus: string;
  block: string;
  floor: string;
  room: string;
  locationLabel: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  assignedDepartment?: string | null;
  reportedBy: string;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  resolvedAt?: string | null;
  /** @nullable */
  resolutionNotes?: string | null;
  /** @nullable */
  duplicateOf?: string | null;
}

export interface IssueInput {
  /** @minLength 12 */
  description: string;
  category?: string;
  campus: string;
  block: string;
  floor: string;
  room: string;
  /** @nullable */
  customLocation?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  reportedBy: string;
}

export interface IssueUpdate {
  status?: string;
  priority?: string;
  category?: string;
  /** @nullable */
  assignedDepartment?: string | null;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  resolutionNotes?: string | null;
}

export interface CreateIssueResult {
  issue: Issue;
  similarIssues: Issue[];
}

export interface IssueAnalysisInput {
  /** @minLength 1 */
  description: string;
}

export interface IssueAnalysis {
  category: string;
  priority: string;
  priorityReason: string;
  confidence: number;
}

export interface StatusHistory {
  id: number;
  issueId: string;
  /** @nullable */
  previousStatus?: string | null;
  newStatus: string;
  changedBy: string;
  timestamp: string;
  /** @nullable */
  note?: string | null;
}

export interface Notification {
  id: number;
  /** @nullable */
  issueId?: string | null;
  message: string;
  read: boolean;
  createdAt: string;
}

export interface DashboardSummary {
  total: number;
  reported: number;
  inProgress: number;
  resolved: number;
  critical: number;
  unreadNotifications: number;
  recentIssues: Issue[];
}

export interface Metric {
  label: string;
  value: number;
}

export interface TrendPoint {
  label: string;
  value: number;
}

export interface Analytics {
  byStatus: Metric[];
  byPriority: Metric[];
  byCategory: Metric[];
  byLocation: Metric[];
  resolutionTrend: TrendPoint[];
  insights: string[];
}

export interface BlockLocation {
  name: string;
  floors: string[];
}

export interface CampusLocation {
  campus: string;
  blocks: BlockLocation[];
}

export type GetIssuesParams = {
mine?: boolean;
search?: string;
category?: string;
priority?: string;
status?: string;
department?: string;
location?: string;
sort?: GetIssuesSort;
};

export type GetIssuesSort = typeof GetIssuesSort[keyof typeof GetIssuesSort];


export const GetIssuesSort = {
  newest: 'newest',
  oldest: 'oldest',
  highest_priority: 'highest_priority',
  recently_updated: 'recently_updated',
} as const;