# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusCare issue reporting and resolution API
 * OpenAPI spec version: 0.1.0
 */
import type { Metric } from './metric';
import type { TrendPoint } from './trendPoint';

export interface Analytics {
  byStatus: Metric[];
  byPriority: Metric[];
  byCategory: Metric[];
  byLocation: Metric[];
  resolutionTrend: TrendPoint[];
  insights: string[];
}