# lib/api-zod/src/generated/types/blockLocation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusCare issue reporting and resolution API
 * OpenAPI spec version: 0.1.0
 */

export interface BlockLocation {
  name: string;
  floors: string[];
}