# lib/api-zod/src/generated/types/campusLocation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusCare issue reporting and resolution API
 * OpenAPI spec version: 0.1.0
 */
import type { BlockLocation } from './blockLocation';

export interface CampusLocation {
  campus: string;
  blocks: BlockLocation[];
}