# lib/api-zod/src/generated/types/createIssueResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusCare issue reporting and resolution API
 * OpenAPI spec version: 0.1.0
 */
import type { Issue } from './issue';

export interface CreateIssueResult {
  issue: Issue;
  similarIssues: Issue[];
}