# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusCare issue reporting and resolution API
 * OpenAPI spec version: 0.1.0
 */
import type { Issue } from './issue';

export interface DashboardSummary {
  total: number;
  reported: number;
  inProgress: number;
  resolved: number;
  critical: number;
  unreadNotifications: number;
  recentIssues: Issue[];
}