# lib/api-zod/src/generated/types/issue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusCare issue reporting and resolution API
 * OpenAPI spec version: 0.1.0
 */

export interface Issue {
  id: number;
  issueId: string;
  description: string;
  shortDescription: string;
  category: string;
  aiCategory: string;
  priority: string;
  aiPriority: string;
  priorityReason: string;
  status: string;
  campus: string;
  block: string;
  floor: string;
  room: string;
  locationLabel: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  assignedDepartment?: string | null;
  reportedBy: string;
  createdAt: Date;
  updatedAt: Date;
  /** @nullable */
  resolvedAt?: Date | null;
  /** @nullable */
  resolutionNotes?: string | null;
  /** @nullable */
  duplicateOf?: string | null;
}