# lib/api-zod/src/generated/types/issueAnalysisInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusCare issue reporting and resolution API
 * OpenAPI spec version: 0.1.0
 */

export interface IssueAnalysisInput {
  /** @minLength 1 */
  description: string;
}