# lib/api-zod/src/generated/types/statusHistory.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusCare issue reporting and resolution API
 * OpenAPI spec version: 0.1.0
 */

export interface StatusHistory {
  id: number;
  issueId: string;
  /** @nullable */
  previousStatus?: string | null;
  newStatus: string;
  changedBy: string;
  timestamp: Date;
  /** @nullable */
  note?: string | null;
}